invoice calculation in progress

+++ db-service/src/main/java/ca/concordia/soen6841/dbservice/repository/SalaryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SalaryRepository extends JpaRepository<Salary, Integer> {
+public interface SalaryRepository extends JpaRepository<Salary, Long> {
 }
